missing documentation that real and float data is rounded
the current documentation is:

>CONCAT implicitly converts all arguments to string types before concatenation. CONCAT implicitly converts null values to empty strings. If CONCAT receives arguments with all NULL values, it will return an empty string of type varchar(1). **The implicit conversion to strings follows the existing rules for data type conversions.** See CAST and CONVERT (Transact-SQL) for more information about data type conversions.

I can't find there something about rounding data when converting float or real into string. This should be more clear in the documentation.

Maybe it should be documented not the way I suggested but in a different way. But something should be in the documentation to avoid wrong usage.